+++ lib/manage-audience/api/manageAudienceClient.ts
@@ -31,7 +31,7 @@ import * as Types from "../../types";
 import { ensureJSON } from "../../utils";
 import { Readable } from "stream";
 
-import HTTPClient from "../../http";
+import HTTPClient from "../../http-axios";
 import { AxiosResponse } from "axios";
 
 // ===============================================
